More work on the MopsProcessor class and its unit test
The list of minimum essential Data Items are now encapsulated in
the DataItemList struct with an associated DataItemListType that
can be either Mandatory or Disjunctive. This change is necessary
because certain Data Items form a logical disjunctive set, meaning
that the presence of any of the Data Items in the set is enough for
the whole set to be compliant with the requirements for the minimum
Data Items.

For the unit tests, two ASTERIX XML files have been created for
verifying the results of the minimum Data Items validity ratio for
Target Reports and Service Messages for both SMR and MLAT.